Change the usage of external `date` to DateTime.
This implies a dependency on the DateTime module, but I believe this is ok
since it is a very used module.
The format string has changed a little, according to DateTime implementation
but it is stable and used anywhere now (I don't see the need for having
two different format string).