Add two new options to pg_restorebinary:
  -E, --exclude-ext EXTNAME    name of an extension to not restore, it can
                               be used multiple time.

  --schema-exists              add an IF NOT EXISTS clause to CREATE SCHEMA

they can be useful to fix extensions import when the creation order have
changed. They an be created manually and excluded from the restoration,
extension's configuration table data are still restored, this just
remove the extension creation statement. If a schema exists before the
restoration, the --schema-exists will prevent the command to fail at
schema creation.